контроллеры

+++ src/main/java/com/example/bus_booking/services/ClientService.java
@@ -1,8 +1,12 @@
 package com.example.bus_booking.services;
 
 import com.example.bus_booking.entities.Client;
+import com.example.bus_booking.entities.RegisterRequest;
 import com.example.bus_booking.repositories.ClientRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.NoSuchElementException;
@@ -13,11 +17,23 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class ClientService {
     private final ClientRepository clientRepository;
+    private final PasswordEncoder passwordEncoder;
 
     public Client createClient(Client client) {
         return clientRepository.save(client);
     }
 
+    public Client findByFirstNameAndLastName(String firstName, String lastName) {
+        return clientRepository.findByFirstNameAndLastName(firstName, lastName);
+    }
+
+
+    public Client updateProfile(String firstName,String lastName, Client updatedClient) {
+        Client client = findByFirstNameAndLastName(firstName, lastName);
+        client.setEmail(updatedClient.getEmail());
+        client.setPhoneNumber(updatedClient.getPhoneNumber());
+        return clientRepository.save(client);
+    }
     public Client getClientById(Long clientId) {
         return clientRepository.findById(clientId)
                 .orElseThrow(() -> new NoSuchElementException("Клиент с id " + clientId + " не найден"));
@@ -46,15 +62,40 @@ public class ClientService {
 
     public void deleteClient(Long clientId) {
         Client client = getClientById(clientId);
-        clientRepository.delete(client);
+        if (client != null) {
+            clientRepository.delete(client);
+        } else {
+            throw new NoSuchElementException("Клиент с ID " + clientId + " не найден");
+        }
     }
+
 
     public boolean isVerified(Long clientId) {
         return clientRepository.existsByIdAndIsVerifiedTrue(clientId);
     }
 
-    public boolean isEmailVerified(Long clientID) {
-        Client client = getClientById(clientID);
+    public boolean isEmailVerified(Long clientId) {
+        Client client = getClientById(clientId);
         return client.isVerifiedMail();
     }
+
+    public ResponseEntity<?> registerClient(RegisterRequest registerRequest) {
+        if (clientRepository.findByEmail(registerRequest.getEmail()) !=null) {
+            return new ResponseEntity<>("Почта уже используется", HttpStatus.BAD_REQUEST);
+        }
+
+        Client client = new Client();
+        client.setFirstName(registerRequest.getName());
+        client.setLastName(registerRequest.getLastName());
+        client.setEmail(registerRequest.getEmail());
+        client.setPassword(passwordEncoder.encode(registerRequest.getPassword()));
+
+        clientRepository.save(client);
+        return new ResponseEntity<>("Регистрация успешна", HttpStatus.OK);
+    }
+
+    public boolean isClientExist(String email) {
+        return clientRepository.existsByEmail(email);
+    }
+
 }
